+++ src/org/soulspace/overarch/adapter/exports/edn.clj
@@ -23,6 +23,7 @@
 (defn export-model
   "Exports the `model` as EDN files."
   [model]
+   
   )
 
 (defmethod exp/export-file :edn