+++ src/org/soulspace/overarch/application/model_repository.clj
@@ -41,7 +41,6 @@
   "Update the accumulator `acc` of the hierarchical model with the element `e`."
   [acc p e]
   (cond
-    ; TODO handle refs
     (el/model-node? e)
     (if (el/child? e p)
       (assoc acc
@@ -51,7 +50,9 @@
                                :id (el/relation-id :parent-of (:id p) (:id e))
                                :from (:id p)
                                :to (:id e)}))
-      (assoc acc :nodes (conj (:nodes acc) (dissoc e :ct))))
+      (assoc acc :nodes (conj (:nodes acc)
+                              ;(dissoc e :ct)
+                              e)))
 
     (el/relation? e)
     (assoc acc :relations (conj (:relations acc) e))
@@ -74,7 +75,7 @@
     ; unhandled element
     :else (do (println "Unhandled:" e) acc)))
 
-(defn relational-model-fn
+(defn ->relational-model
   "Step function for the conversion of the hierachical input model into a relational model of nodes, relations and views."
   ([] [{:nodes #{}
        :relations #{}
@@ -90,7 +91,7 @@
 (defn build-relational-model
   "Builds a relational working model from the hierarchical inpur model."
   [coll]
-  (let [relational (model/traverse relational-model-fn coll)]
+  (let [relational (model/traverse ->relational-model coll)]
     ; TODO add registries
     ; :id->element
     ; :id->referred-relations