+++ src/org/soulspace/overarch/domain/views/state_machine_view.clj
@@ -4,11 +4,11 @@
 
 (defmethod view/render-model-node? :state-machine-view
   [view e]
-  (contains? e/state-machine-types (:el e)))
+  (contains? e/state-machine-view-types (:el e)))
 
 (defmethod view/include-content? :state-machine-view
   [view e]
-  (contains? e/state-machine-types (:el e)))
+  (contains? e/state-machine-view-types (:el e)))
 
 (defmethod view/render-relation-node? :state-machine-view
   [view e]